Add FPS counter (doesn't work properly in over-30)
For some reason it doesn't update in over-30-mode. But otherwise it
works fine in 30-mode.

Also, semi-unfinished. I was planning on adding a keybind toggle or
something, and also making it so the measurements are averaged out so
every time it updates you don't get weird outlier numbers, but this will
do for now.