Add artifact uploads to CI
This adds the uploading of artifacts, most notably the resulting
compiled binary, after the automatic Continuous Integration workflow is
run for the default version and the Make & Play version.

Upstream has clearly stated they do not wish to have their CI artifacts
uploaded (see issue 62 in the upstream repository, which I will not
directly link to because otherwise it'll say that I linked to it on that
issue). So I'm doing what I recommended in that issue, which is to just
fork the repo and add CI myself.