Allow using internal commands in simplified with brackets
This allows the use of internal commands in simplified, via "internal
mode blocks" which start with and end with square brackets ('[' and
']').

This doesn't add any functionality that wasn't already accessible -
people could already use internal commands by using one of two exploit
paths. However, either exploit path isn't really convenient to use.
These square brackets are just only there for convenience. You don't
even need the closing bracket - you can put an opening bracket at the
top of the script, and the rest of the script will be in internal mode!

Why not simply make it so the parser doesn't throw away internal
commands? Well, that results in an issue with certain existing internal
scripts, which are one-line internal commands that get jumped to from an
iftrinkets() - directly making every script internal will result in this
one-line command being executed twice. This is because the game normally
throws it out the first time, but then falls victim to the last-line
evaluation bug (part of the internal scripting exploit path) and
ends up executing the line as internal - if the game didn't throw out
the line in the first place, then it ends up executing the script twice.

For now, I've decided to go with the opt-in method, instead of doing
internal-by-default - resolving the double execution bug seems to get
pretty complicated very quickly. (VVVVVV: Community Edition (which is
dead) resolved it by simply only allowing internal commands in
simplified if it was a VCE level, which is not something that I want to
do here.)