Add dead bodies
Last night (the night of December 4th, 2020), I was laying in bed with
Coral on a Discord voice call and she was thinking about how checkpoints
worked. (Like, an in-lore explanation, not how it works in the actual
game.) Then she suggested at one point, what if there was a mod for
VVVVVV where you left behind a dead body every time you died.

So behold, this is that mod. What happens every time you die is the game
keeps track of where exactly you died, along with your sprite at the
time of death, and shoves it into an array. Then the game simply draws
that array on the screen, using color 1 (which is the color of death).
In the main game, it's drawn above all other non-humanoids, but in
custom levels it's drawn before every other entity.

Supercrewmates (the Intermission 1 companion) also create dead bodies.

Dying as VVVVVV-Man doesn't show a VVVVVV-Man-sized body, but the
playtesting ghosts feature doesn't account for VVVVVV-Man either, so.
Possibly playtesting ghosts and dead bodies should just use the same
entity-drawing function as normal entities do, but that's a bit
complicated. Or at the very least, some code should be refactored and
de-duplicated there to easily enable that without having to copy-paste
code.

Dead bodies are cleared every time you quit to the menu, or exit
playtesting (so, every time scriptclass::hardreset() gets called).

Whenever a dead body is added, there's also a check so the game doesn't
add a duplicate dead body if there's already a dead body that exists
with the same exact position and drawframe.