Implement initial support for replays and LTM converter script
This implements initial support for replays, as well as initial support
for an LTM converter script.

The initial support is that the game can write replays, but it cannot
play them back. Instead, I'm testing replay functionality by converting
the inputs to LTM (libTAS movie) format, and pasting them in to a stub
LTM file. So far, the inputs seem to be captured correctly because they
play back correctly when I play back the LTM.

I've decided that the input format is 1 frame per line listing every
action taken on a gameplay frame - that is, when gameinput() is called.
To maintain determinism, I'm also recording somewhat external actions
like enabling/disabling invincibility, even though they're not strictly
done in gameinput().

There is a game option to enable replays when gameplay is started. It is
disabled by default.

Replays are currently only supported for the default mode of the main
game, and custom levels. Time trials, No Death Mode, etc. are not yet
supported.

Replays are saved in a new replays/ directory in the VVVVVV directory.

Save files are supported by automatically copying the loaded save file
to the replays directory, so you don't have to do it yourself and have a
record of where to start off in order to get the replay to work.

Replay files are appended to every frame, and not written all at once,
in case the game closes unexpectedly. This may or may not be good for
performance. I have a super-fast SSD so personally I don't care.

This also adds an option to delete all replay files.

Console output (at INFO level) is given for opening a replay file,
copying a save file for a reply, and closing a replay file.

To indicate that a replay will be recorded before loading into gameplay,
this draws the text "Replays are on" in the corner of the screen before
any point where the player could load into gameplay that starts a replay
recording. That is:

- On the title screen
- In the play menu (where you can select "continue" after selecting
  "play")
- In the continue menu (where you choose between quicksave or telesave)
- In the custom levels list
- In the custom level quicksave menu (use quicksave or start from
  beginning)